+++ src/net/game/item/stuff/Stuff.java
@@ -50,7 +50,7 @@ public class Stuff extends Item {
 		this.mana = stuff.mana;
 	}
 	
-	public Stuff(StuffType type, ClassType[] classType, String sprite_id, int id, String name, int quality, GemColor color1, GemColor color2, GemColor color3, GemBonusType gemBonusType, int gemBonusValue, int level, Wear wear, int critical, int strength, int stamina, int armor, int mana, int sellPrice) {
+	public Stuff(StuffType type, ClassType[] classType, String sprite_id, int id, String name, byte quality, GemColor color1, GemColor color2, GemColor color3, GemBonusType gemBonusType, int gemBonusValue, int level, Wear wear, int critical, int strength, int stamina, int armor, int mana, int sellPrice) {
 		super(id, sprite_id, ItemType.STUFF, name, quality, sellPrice, 1, 1);
 		this.gemBonusValue = gemBonusValue;
 		this.gemBonusType = gemBonusType;
@@ -86,7 +86,7 @@ public class Stuff extends Item {
 		this.mana = weapon.mana;
 	}
 	
-	public Stuff(int id, String name, String sprite_id, ClassType[] classType, WeaponType weaponType, WeaponSlot weaponSlot, int quality, GemColor color1, GemColor color2, GemColor color3, GemBonusType gemBonusType, int gemBonusValue, int level, int armor, int stamina, int mana, int critical, int strength, int sellPrice) {
+	public Stuff(int id, String name, String sprite_id, ClassType[] classType, WeaponType weaponType, WeaponSlot weaponSlot, byte quality, GemColor color1, GemColor color2, GemColor color3, GemBonusType gemBonusType, int gemBonusValue, int level, int armor, int stamina, int mana, int critical, int strength, int sellPrice) {
 		super(id, sprite_id, ItemType.WEAPON, name, quality, sellPrice, 1, 1);
 		this.gemBonusValue = gemBonusValue;
 		this.gemBonusType = gemBonusType;
